// types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      events: {
        Row: {
          background_image_url: string | null
          couple_id: string
          created_at: string | null
          id: string
          settings: Json | null
          slug: string
          title: string
        }
        Insert: {
          background_image_url?: string | null
          couple_id: string
          created_at?: string | null
          id?: string
          settings?: Json | null
          slug: string
          title: string
        }
        Update: {
          background_image_url?: string | null
          couple_id?: string
          created_at?: string | null
          id?: string
          settings?: Json | null
          slug?: string
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "events_couple_id_fkey"
            columns: ["couple_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["couple_id"]
          },
        ]
      }
      submissions: {
        Row: {
          content: string | null
          created_at: string | null
          event_id: string
          id: string
          is_favorite: boolean | null
          moderated: boolean | null
          sender_contact: string | null
          sender_name: string
          storage_meta: Json | null
          storage_path: string | null
          type: string
        }
        Insert: {
          content?: string | null
          created_at?: string | null
          event_id: string
          id?: string
          is_favorite?: boolean | null
          moderated?: boolean | null
          sender_contact?: string | null
          sender_name: string
          storage_meta?: Json | null
          storage_path?: string | null
          type: string
        }
        Update: {
          content?: string | null
          created_at?: string | null
          event_id?: string
          id?: string
          is_favorite?: boolean | null
          moderated?: boolean | null
          sender_contact?: string | null
          sender_name?: string
          storage_meta?: Json | null
          storage_path?: string | null
          type?: string
        }
        Relationships: [
          {
            foreignKeyName: "submissions_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          },
        ]
      }
      users: {
        Row: {
          couple_id: string
          created_at: string | null
          email: string
          id: string
          role: string | null
        }
        Insert: {
          couple_id: string
          created_at?: string | null
          email: string
          id?: string
          role?: string | null
        }
        Update: {
          couple_id?: string
          created_at?: string | null
          email?: string
          id?: string
          role?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never